WIP  - feature #354 Update Integration to handle files from API   - feature #354 Update Integration to handle files from API

+++ gendox-core-api/src/main/java/dev/ctrlspace/gendox/gendoxcoreapi/services/integrations/ApiIntegrationUpdateService.java
@@ -38,9 +38,7 @@ public class ApiIntegrationUpdateService implements IntegrationUpdateService {
     }
 
     String baseUrl = "https://test.dma.com.gr/wp-json";
-    String apiKey = "k9gz4TUVuUtEdKZnMNWZk7fg4B63bquX";
-
-
+    String apiKey = "eyJhbGciOiJSUzI1NiIsInR5cCIgOiAiSldUIiwia2lkIiA6ICJDR0NpMlNTX2lQNkdGYTBKQmVqRjAxYzNpcDBTdm43d2FLMGNYQnJHR19RIn0.eyJleHAiOjE3MzI0MDE1NjgsImlhdCI6MTczMjM1ODM2OCwiYXV0aF90aW1lIjoxNzMyMzU4MzY3LCJqdGkiOiJkNDU4MjliZi03YzQ0LTRlMmQtYTZkNS1hOGMxNzk4ZGQzNTUiLCJpc3MiOiJodHRwczovL2Rldi5nZW5kb3guY3RybHNwYWNlLmRldi9pZHAvcmVhbG1zL2dlbmRveC1pZHAtZGV2IiwiYXVkIjoiYWNjb3VudCIsInN1YiI6IjEyYjYwNGQ4LTk1OTktNDk3ZS05ZDA4LTAwYjdkZmQ5ZmVkMiIsInR5cCI6IkJlYXJlciIsImF6cCI6ImdlbmRveC1wa2NlLXB1YmxpYy1jbGllbnQtZGV2Iiwic2lkIjoiZTFhM2Y5NDgtZjkzNy00N2NkLWI2M2QtZDRhZjk4OTZhZGY3IiwiYWNyIjoiMSIsImFsbG93ZWQtb3JpZ2lucyI6WyJodHRwczovL2Rldi5nZW5kb3guY3RybHNwYWNlLmRldiJdLCJyZWFsbV9hY2Nlc3MiOnsicm9sZXMiOlsiZGVmYXVsdC1yb2xlcy1nZW5kb3gtaWRwLWRldiIsIm9mZmxpbmVfYWNjZXNzIiwidW1hX2F1dGhvcml6YXRpb24iXX0sInJlc291cmNlX2FjY2VzcyI6eyJhY2NvdW50Ijp7InJvbGVzIjpbIm1hbmFnZS1hY2NvdW50IiwibWFuYWdlLWFjY291bnQtbGlua3MiLCJ2aWV3LXByb2ZpbGUiXX19LCJzY29wZSI6Im9wZW5pZCBlbWFpbCBwcm9maWxlIiwiZW1haWxfdmVyaWZpZWQiOnRydWUsIm5hbWUiOiJBY2NvdW50IG9uZSBUZXN0QWNjb3VudCAiLCJwcmVmZXJyZWRfdXNlcm5hbWUiOiJzZWxvbW81MDA5QGV4d2VtZS5jb20iLCJnaXZlbl9uYW1lIjoiQWNjb3VudCBvbmUiLCJmYW1pbHlfbmFtZSI6IlRlc3RBY2NvdW50ICIsImVtYWlsIjoic2Vsb21vNTAwOUBleHdlbWUuY29tIn0.g5q4vjxCo7hv-l_NIEKxetNEvny7nlzmmjWHHZDfXbFekcDLaB9qFLM9oZYDvU2FwmT9rM3lNKORp6nnVw2x6DFfecWo1m0of6Ov5md04onu5OdSvDhYHzcwjROXGX5-8zNJxi65ZIpCBXWJmZdKhAs0pIWdYB6USGFzWkfqUpqtw4UnypUgqckDFVOYLLswfY8fpGYiJnimPvwCy-DLPpVS6Ic97UgwrGSTcievQOielLUokyGGzS6igSJLw93y24ZAwEytt8whPLAeuPqPRUoNrNN2leZK2tAzrpHsso5cPb6SQ-mSGFFbHDBcbdC_1RcGbm7qsHQu_jxv37K2_Q";
     @Override
     @Transactional(value = Transactional.TxType.REQUIRES_NEW)
     public Map<ProjectIntegrationDTO, List<IntegratedFileDTO>> checkForUpdates(Integration integration) throws GendoxException {
@@ -62,10 +60,11 @@ public class ApiIntegrationUpdateService implements IntegrationUpdateService {
         logger.debug("Docs to Delete (DocumentInstance IDs): {}", docsToDelete.size());
 
         //DELETE files in docsToDelete
-        documentService.deleteAllDocumentInstances(docsToDelete);
+        if (!docsToDelete.isEmpty()) {
+            documentService.deleteAllDocumentInstances(docsToDelete);
+        }
 
         //create from docsToCreate and docsToUpdate a map of Map<project_id, ContentIDDto>
-
         Map<UUID, List<TempIntegrationFileCheck>> projectContentMap = new HashMap<>();
         // populate the above map
         Stream.of(docsToUpdate.stream(), docsToCreate.stream())
@@ -78,6 +77,9 @@ public class ApiIntegrationUpdateService implements IntegrationUpdateService {
                     projectContentMap.get(projectId).add(tempFile);
                 });
 
+        if (projectContentMap.isEmpty()) {
+            tempIntegrationFileCheckService.deleteTempIntegrationFileChecksByIntegrationId(integration.getId());
+        }
 
         organizationMap = createMap(projectContentMap, integration);
